+++ server/pom.xml
@@ -124,6 +124,12 @@
             <scope>test</scope>
             <version>1.1.3</version>
         </dependency>
+    
+      <dependency>
+        <groupId>com.google.collections</groupId>
+        <artifactId>google-collections</artifactId>
+        <version>1.0</version>
+      </dependency>
     </dependencies>
 
     <build>
