+++ server/src/main/java/com/google/finapp/ServerMain.java
@@ -15,6 +15,8 @@
 package com.google.finapp;
 
 import com.beust.jcommander.JCommander;
+import com.beust.jcommander.Parameter;
+import com.beust.jcommander.Parameters;
 import com.google.cloud.spanner.DatabaseClient;
 import com.google.cloud.spanner.DatabaseId;
 import com.google.cloud.spanner.Spanner;
@@ -53,5 +55,31 @@ public final class ServerMain {
         spanner.getDatabaseClient(
             DatabaseId.of(spannerProjectId, spannerInstanceId, spannerDatabaseId));
     return new SpannerDaoImpl(client);
+  }
+
+  @Parameters(separators = "=")
+  private static class Args {
+    @Parameter(names = {"--port", "-p"})
+    int port = 8080;
+
+    @Parameter(names = {"--spanner_host"})
+    String spannerHost = "spanner.googleapis.com";
+
+    @Parameter(names = {"--spanner_port"})
+    int spannerPort = 443;
+
+    @Parameter(names = {"--spanner_project_id"})
+    String spannerProjectId;
+
+    @Parameter(names = {"--spanner_instance_id"})
+    String spannerInstanceId;
+
+    @Parameter(names = {"--spanner_database_id"})
+    String spannerDatabaseId;
+
+    @Parameter(
+        names = {"--spanner_use_jdbc"},
+        arity = 0)
+    boolean spannerUseJdbc = false;
   }
 }
